requirements.txt
asgiref==3.8.1
attrs==24.3.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
defusedxml==0.8.0rc2
Django==5.1.4
django-filter==24.3
django-redis==5.4.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-nested-routers==0.94.1
drf-spectacular==0.28.0
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
kavenegar==1.1.2
oauthlib==3.2.2
pillow==11.0.0
pycparser==2.22
PyJWT==2.10.1
python-slugify==8.0.4
python3-openid==3.2.0
PyYAML==6.0.2
redis==5.2.1
referencing==0.35.1
requests==2.32.3
requests-oauthlib==2.0.0
rpds-py==0.22.3
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
text-unidecode==1.3
uritemplate==4.1.1
urllib3==2.2.3
